Fixed user mentions for extensions
When run from an extension, message.mentions now returns an array of
Member objects. For some bizarre reason Eris returns the mentions array
by user ID in ascending order (totally illogical, in my opinions), so
through cumbersome means I've reordered the array in order of the user
mentioned in the message. The snag is that multiple mentions of the same
user will be disregarded (but why would anyone do that?)

Probably not the most elegant solution possible, but it solves the
problem of mentions not working, puts mentions in order and passes
additional user information about mentioned users, such as AvatarURL,
status, etc. to extensions.