Fixed bug in time <user> where the user has not set a location
If time <user> is run but the user has not set a profile location, an
error results as the message suffix is treated as a location. Fixed by
adding an additional parameter to getTime() that is only used where the
user option is provided and that user has no location set.