Weather unit added to profile
The default temperature unit for the weather command was Fahrenheit, but
this is uncommon in most of the world (which uses Celsius). Setting to F
or C is problematic because whoever doesn't use the default is forced to
define the temperature unit every time they use the weather command.

For example, if a user wants to see the weather in Paris in Celsius they
must type `weather Paris C`. If that user then sets their profile
location to Paris, they can simply run `weather`. But this convenience
is undermined by the fact that `weather` will show them the weather in
Paris in **Fahrenheit**. So despite setting their profile location, they
must run `weather Paris C` every time.

My initial thought was to simply set the default to C, but this causes
the inverse problem for those using Fahrenheit (although we know who
[they
are](https://qph.ec.quoracdn.net/main-qimg-4b106c00694e6ceefd236d40928d9a3f-c)).
So I've instead added a new key to the user profile schema: weatherunit.
A user can now PM `profile weatherunit|<fahrenheit or celsius>` and the
weather command will in future respond according to their preference,
defaulting to Celsius if th weatherunit key has not been set.