Fixed numerous issues with giveaway
1) The giveaway command was ignoring any duration other than the default
(per issue #70).

2) Giveaways were not ending by themselves at the conclusion of the period
set. They could only be ended manually by the creator via PM.

This was because Giveaways.end(), when triggered on a setTimeout() by
Giveaways.start() was passing the channelDocument as it was when the
giveaway began (thus meaning the participant_ids array was always
empty). On the other hand, when ended manually, the channelDocument
that was passed to Giveaways.end() was up to date.

Giveaways.end() has been updated to use the live channelDocument, and
consequently the problem has been eliminiated.

3) The success message PM'd to the creator would always say a winner
could not be chosen, even where a winner had been chosen. Also, the
creator would only be PM'd the results of their giveaway when they
forcefully ended it by PM. I have resolved both these issues by
removing the PM from the manual end awaitMessage() in giveaways.js
and altering the Giveaways.end() function to always PM the creator
at the conclusion of their giveaway, regardless of how it ended.
The PM also now correctly identifies the winning user.