Added options parameter to Guild.createRole() (extensions)
The parameter remains optional, but if we don't pass it to the function, we then have
to use the callback and run Role.edit() on it to specify options for the role we've
created. It is much more efficient to be able to specify options when creating a role
in the first place.