Added missing violator_role_id
The violator_role_id should be defined in the serverConfig schema as an
empty String with a default "". Because it isn't, the select field
entitled "Role assigned to violators" in the web interface (for each
filter) can never save, because the post handling for filters only
updates *existing* fields. Updating the serverConfig schema is no good
for those with existing GAB installs, and this solution works just fine
for everyone.