Fixed numerous issues with filters
Spam handling was flawed because the spam document was not being saved
after each message. Additionally, the delayed deletion in the setTimeout
would not work correctly because by the time it executed the object it
was trying to remove had been updated and so was no longer the 'same'
object. This is resolved by initializing a new db.servers.findOne()
within the setTimeout.

Instances of bot.handleViolation have also been updated to check if a
message author is already a member of the violator role (if defined), to
avoid trying to add the member to the role twice. If the member is
already a member of the role, this produces an error 400 on the Discord
API and is simply unnecessary.