Added option to filter everyone/here in mentions
The mention filter did not include everyone/here in the mention
sensitivity count. I've added this in, but made it optional.

Web interface and schema have been updated to support this new option
under the mention filter. Schema has also been updated to specify
violator_role_id with an empty default, so that changes to the violator
role for each filter save correctly in a post from the web interface.