Translation now uses MS Azure Cognitive Services
Following the retirement of Microsoft DataMarket at the end of March
2017, the translation command and auto-translate feature will no longer
work. Additionally, those new to AB/GAB have not been able to register
API keys on DataMarket as this was closed in December 2016.

Given that Microsoft provides an excellent translation API for free
(with up to 2 million characters), it seems sensible to continue using
Microsoft (as opposed to other translation APIs).

The translation command and auto-translate feature are both now fully
working in GAB, using the new MS Cognitive Services API.

Existing GAB hosters will need to:

1) obtain a Cognitive Services API key (this is free; I will post
instructions on how to do this in due course).
2) edit their Configuration/auth.json to specify the new API key (under
the field: `microsoft_cs_key` [the old microsoft_client_id and
microsoft_client_secret fields can be removed]).
3) run `npm update` to install the MsTranslator node module.