Avoid translating from English in Auto-Translate
Auto-translate allows for a particular user's messages to be translated
automatically from a specific language into English. But if that user is
typing messages some of which are in English, translation for those
English messages is not necessary.

While adding a language detection request on each message means another
API call, it ultimately saves on translation usage. The free plan with
Microsoft provides a quota of up to 2 million characters translated per
month. There is no point in wasting usage by translating messages that
are already in English. The *detect* request does not use any of the
translation character quota and there is no limit on detection usage.