[SharedCommands/Events] Shared command system :tada:
Shared commands are the most basic commands. They don't expect much info.
It's at the developers choice whether to make them a separate command (like eval) or make a public and private command which in turn, will run the shared command (and not defining it in commands.js in that case)
Things are still experimental, but they work!