Update maintainer perm defaults
With the evaluate permission, maintainers can "elevate" their perms to any level. Free execution of JavaScript enables users to have full control over the entirety of GAB, so it should be defaulted to The Host only. The shutdown commands are "shortcuts" of eval commands that sudo maintainers can have access to without undermining the permission system.