Remove "?svrid" in extension builder form as well
After careful consideration I came to the conclusion that half of the code is outdated anyway (isGallery has been removed already) but at least with this fix it works again.
The last ?svrid left is the one in /userlist which is okay because the WebServer path handler manually asks for that param.