Fix winston throwing a circular JSON error on e.g. Snekfetch errors (#258)
When logging stuff to the JSON log file winston shoves the output throuh JSON.stringify by default. The problem with this is that errors thrown by for example Snekfetch contain circular references. The error contains the response object which contains the request object which contains the response again. So in the end logging such an error throws an error itself.
Fix that by using circular-json#stringify instead of the default one.