:ambulance: Fix statsCollector not finishing
This also fixes the READY event not finishing and thus commands never being loaded.

We simply ignore the falsy value rather than trying to handle it, as it's not this code's purpose to handle false memberDocuments and nor do we expect it to.

GAB can not operate if broken subDocuments are present in a serverDocument. While it is currently unknown what causes these broken subDocuments or if it's already fixed, we'll be on the lookout for code that can cause these subDocuments to be pushed (eventhough all db operations should use mongoose models, and mongoose validation is supposed to prevent exactly this from happening).