Fixed TypeError
Basically, I found that if you cause/execute an error that is undefined/null, when the bot tries to send err.stack, it will result in a TypeError because err is undefined, and so it can't read the stack property. This is a very simple fix.

You can replicate this error yourself by just trying to execute Promise.reject().